Factor out Date::Infinity to a separate file
